requirements.txt
flask
openai-whisper
torch
deep-translator
werkzeug
pydub
